requirements.txt
amqp==5.3.1
asgiref==3.11.1
asttokens==3.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
decorator==5.2.1
Django==6.0.2
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-timezone-field==7.2.1
django-user-agents==0.4.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
executing==2.2.1
greenlet==3.3.2
hiredis==3.3.0
idna==3.11
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.6.2
matplotlib-inline==0.2.1
mysqlclient==2.2.7
packaging==26.0
parso==0.8.5
pgsql==2.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.9.0
pyotp==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2023.3
redis==5.3.0
requests==2.32.0
six==1.17.0
SQLAlchemy==2.0.47
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==202602
urllib3==2.6.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.6.0
